requirements.txt
discord.py
python-dotenv
duckduckgo-search
selenium
ollama
requests
mcp
